refactor change: all management pages NamePage.jsx -> MNamePage.jsx to idicate a Management Page to avoid conflict
