+++ src/App/Management/Subject/MSubjectsPage.css
@@ -1,0 +1,3 @@
+.MSubjectPage .field-list{
+    margin-top: 10px;
+}
