--- src/App/Management/Subject/SubjectsPage.css
@@ -1,3 +1,0 @@
-.SubjectPage .field-list{
-    margin-top: 10px;
-}
